+++ NpgsqlRest/Interfaces.cs
@@ -97,7 +97,7 @@ public interface IRoutineSource
     /// <summary>
     /// Comments mode for the current routine source.
     /// </summary>
-    CommentsMode? CommentsMode { get => null; }
+    CommentsMode? CommentsMode { get; set; }
 
     /// <summary>
     /// Yield all routines with the formatters from the current source.
@@ -109,6 +109,47 @@ public interface IRoutineSource
     /// <summary>
     /// SQL Query that returns data source.
     /// When it doesn't contain any blanks, it is interpreted as a function name.
+    /// Set to NULL to use default source query.
     /// </summary>
-    string Query { get; set; }
+    string? Query { get; set; }
+    
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string? SchemaSimilarTo { get; set; }
+
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string? SchemaNotSimilarTo { get; set; }
+
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string[]? IncludeSchemas { get; set; }
+
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string[]? ExcludeSchemas { get; set; }
+
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string? NameSimilarTo { get; set; }
+
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string? NameNotSimilarTo { get; set; }
+
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string[]? IncludeNames { get; set; }
+
+    /// <summary>
+    /// Query parameter
+    /// </summary>
+    string[]? ExcludeNames { get; set; }
 }